[refactor] GGBE-101 delete entity in admin
