+++ src/main/java/com/gg/server/admin/penalty/dto/PenaltyUserResponseDto.java
@@ -1,6 +1,6 @@
 package com.gg.server.admin.penalty.dto;
 
-import com.gg.server.admin.penalty.data.RedisPenaltyUser;
+import com.gg.server.domain.penalty.data.RedisPenaltyUser;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
